refactor(dashboard-v2): extract add-variable buttons into components
